+++ src/pages/Dashboard.js
@@ -1,5 +1,6 @@
 import React, { useEffect } from "react"
 import { connect } from "react-redux"
+import Layout from '../components/Layout'
 import Spinner from "../components/UI/Spinner/Spinner"
 
 const Dashboard = ({ state, dispatch }) => {
@@ -7,7 +8,11 @@ const Dashboard = ({ state, dispatch }) => {
         // we send our requests here via 'dispatch'
     }, [])
 
-    return state.loading ? <Spinner /> : <h1>Dashboard</h1>
+    return (
+        <Layout >
+            {state.loading ? <Spinner /> : <h1>Dashboard</h1>}
+        </Layout>
+    )
 }
 
 export default connect(state => ({
